requirements.txt
pyaudio
SpeechRecognition[whisper-api]
wave
threading
requests
pygame
openai-whisper
time
spotipy
pyppeteer = 1.0.0
google-api-python-client
flask